Adding optional OAuth 2.0 support
In order to reduce moderation burden for people running their own instances, this adds the option for an administrator to configure an OAuth domain for anyone who tries to create a new event, import events, or create a new event group. The idea is that if I am running a Mastodon server at https://social.example, then I can limit event creation to people who have a https://social.example account. This means that I am moderating the people I have already agreed to moderate, and removes the chance of random internet people creating a "free gold watch" spam event on my server, or worse. Editing via the special edit link still works as normal, so you'd need an account to create an event, but you could share the edit link with anyone and they could co-moderate with you. Comments and RSVPs remain open to anyone who knows about the event as well.